requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic==2.5.0
pydantic-settings==2.11.0
pytest==8.4.2
crewai==0.1.0
langchain==0.1.0
sentence-transformers==2.2.2
chromadb==0.4.18
scikit-learn==1.3.2
httpx==0.25.2